pyproject.toml
[tool.poetry]
name = "scalesops"
version = "0.1.0"
description = ""
authors = ["Ahmad Sadeghi <ahmad.81.sadeghi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.9"
bs4 = "^0.0.1"
pillow = "^10.2.0"
requests = "^2.31.0"
mock = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
